+++ Scripts/toloadhub.lua
@@ -10,17 +10,6 @@
     Features include automatic loading from SimBrief, random passenger generation, and real-time loading.
 
     TODO: Missing max Fuel and Cargo for A339 and A346
-    TODO: DOORS:
-    Clicca se DOOR sempe e solo manuali se no:
-    Before staqrt se:
-    Airstair vs Jetbridg
-    Airstair:
-    A319 A20N A321 A21N:COme ora, selezioni o L1 o L4 IN DEBOARDING NEVER L4
-    A346 and A339: L1, L2, L4 door should open. (capire se la l2 che indice ha) da opzioni preselezionare se L1 e L4 e anche L2 (la 6??)
-
-    Jetbridg:
-    A319 A20N A321 A21N: SEMPRE e solo L1 anche per DEBOARDIN
-    A346 and A339: Dare una opzione che mostra L1, L2 e L4 e selzioni quello che vuoi.
     License: MIT License
 --]]
 
@@ -31,10 +20,14 @@ if not valid_plane_icao[PLANE_ICAO] then
     return
 end
 
+--    Airstair vs Jetbridge--
+--    Jetbridge:
+--    A319 A20N: Only L1
+
 -- == CONFIGURATION DEFAULT VARIABLES ==
 local toLoadHub = {
     title = "ToLoadHUB",
-    version = "0.11.1",
+    version = "0.12.0",
     file = "toloadhub.ini" ,
     visible_main = false,
     visible_settings = false,
@@ -54,6 +47,7 @@ local toLoadHub = {
     kgPerUnit = 25,
     first_init = false,
     is_lbs = false,
+    is_jetbridge = false,
     unitLabel = "KGS",
     unitTLabel = "T",
     phases = {
@@ -97,11 +91,17 @@ local toLoadHub = {
     deboarding_cargo_sound_played = false,
     setWeightCommand = false,
     full_deboard_sound = false,
+    chocks_off_time = os.date("%H:%M:%S"),
+    chocks_on_time = os.date("%H:%M:%S"),
+    chocks_off_set = false,
+    chocks_on_set = false,
     hoppie = {
         loadsheet_sent = false,
         loadsheet_sending = false,
         loadsheet_preliminary_ready = false,
         loadsheet_preliminary_sent = false,
+        loadsheet_chocks_off_sent = false,
+        loadsheet_chocks_on_sent = false,
         loadsheet_check = os.time(),
     },
     simbrief = {
@@ -135,7 +135,8 @@ local toLoadHub = {
         hoppie = {
             secret = "",
             enable_loadsheet = true,
-            preliminary_loadsheet = false
+            preliminary_loadsheet = true,
+            chocks_loadsheet = true,
         },
         door = {
             close_boarding = true,
@@ -149,7 +150,7 @@ local toLoadHub = {
 local loadsheetStructure = {
     new = function(self)
         local obj = {
-            isFinal = false,
+            typeL = 0, -- 0 = preliminary, 1 = final, 2 = chocks off, 3 = chocks on
             warning = "",
             labelText = "",
             flt_no = "",
@@ -360,6 +361,16 @@ local function setIscsTemporarySimbrief()
     toLoadHub_FwdCargo_XP = toLoadHub.cargo_fwd
 end
 
+local function isAllEngineOff()
+    local engine = dataref_table("sim/flightmodel/engine/ENGN_running")
+    local all_zero = true
+
+    for _, value in ipairs(my_table) do
+        if value > 0 then all_zero = false; break end
+    end
+    return all_zero
+end
+
 local function setAirplaneNumbers()
     if PLANE_ICAO == "A319" then
         toLoadHub.max_passenger = 145
@@ -405,7 +416,12 @@ local function resetAirplaneParameters()
     toLoadHub.wait_until_speak = os.time()
     toLoadHub.setWeightTime = os.time()
     toLoadHub.next_ready_to_start_check = os.time()
+    toLoadHub.chocks_off_time = os.date("%H:%M:%S")
+    toLoadHub.chocks_on_time = os.date("%H:%M:%S")
+    toLoadHub.chocks_off_set = false
+    toLoadHub.chocks_on_set = false
     toLoadHub.what_to_speak = nil
+    toLoadHub.is_jetbridge = false
     toLoadHub.boarding_sound_played = false
     toLoadHub.deboarding_sound_played = false
     toLoadHub.boarding_cargo_sound_played = false
@@ -542,6 +558,7 @@ end
 local function openDoors(boarding)
     local setVal = boarding and toLoadHub.settings.door.open_boarding or toLoadHub.settings.door.open_deboarding
     toLoadHub_Doors_1 = 2
+    if toLoadHub.is_jetbridge and PLANE_ICAO == "A319" or PLANE_ICAO == "A20N" then return end
     if setVal > 1 then
         toLoadHub_Doors_2 = 2
         if PLANE_ICAO == "A321" or PLANE_ICAO == "A21N" or PLANE_ICAO == "A346" or PLANE_ICAO == "A339" then
@@ -639,17 +656,30 @@ local function sendLoadsheetToToliss(data)
     end
 
     toLoadHub.hoppie.loadsheet_sending = true
-
-    local loadSheetContent = "/data2/313//NE/" .. table.concat({
-        "Loadsheet " .. data.labelText .. " " .. os.date("%H:%M"),
-        formatRowLoadSheet("ZFW",  data.zfw, 9),
-        formatRowLoadSheet("ZFWCG", data.zfwcg, 9),
-        formatRowLoadSheet("GWCG", data.gwcg, 9),
-        formatRowLoadSheet("F.BLK", data.f_blk, 9),
-    }, "\n")
-    if data.warning ~= "" then
-        loadSheetContent = loadSheetContent .. "\n" .. formatRowLoadSheet("@WARN!@ F.BLK EXP.", data.warning, 22)
+    local loadSheetContent = ""
+    if data.typeL < 2 then
+        loadSheetContent = "/data2/313//NE/" .. table.concat({
+            "Loadsheet " .. data.labelText .. " " .. os.date("%H:%M"),
+            formatRowLoadSheet("ZFW",  data.zfw, 9),
+            formatRowLoadSheet("ZFWCG", data.zfwcg, 9),
+            formatRowLoadSheet("GWCG", data.gwcg, 9),
+            formatRowLoadSheet("F.BLK", data.f_blk, 9),
+        }, "\n")
+        if data.warning ~= "" then
+            loadSheetContent = loadSheetContent .. "\n" .. formatRowLoadSheet("@WARN!@ F.BLK EXP.", data.warning, 22)
+        end
+    elseif data.typeL == 2 then
+        loadSheetContent = "/data2/323//NE/" .. table.concat({
+            "Loadsheet " .. data.labelText .. " " .. os.date("%H:%M"),
+            data.flt_no .. " Chocks Off time " .. toLoadHub.chocks_off_time
+        }, "\n")
+    elseif data.typeL == 3 then
+        loadSheetContent = "/data2/333//NE/" .. table.concat({
+            "Loadsheet " .. data.labelText .. " " .. os.date("%H:%M"),
+            data.flt_no .. " Chocks On time " .. toLoadHub.chocks_on_time
+        }, "\n")
     end
+
 
     debug(string.format("[%s] Hoppie flt_no %s.", toLoadHub.title, tostring(data.flt_no)))
 
@@ -672,8 +702,11 @@ local function sendLoadsheetToToliss(data)
         source = ltn12.source.string(payload),
     }
     debug(string.format("[%s] Hoppie returning code %s.", toLoadHub.title, tostring(code)))
-    if code == 200 and data.isFinal then toLoadHub.hoppie.loadsheet_sent = true end
-    if code == 200 and not data.isFinal then toLoadHub.hoppie.loadsheet_preliminary_sent = true end
+    if code == 200 and data.typeL == 0  then toLoadHub.hoppie.loadsheet_sent = true end
+    if code == 200 and data.typeL == 1 then toLoadHub.hoppie.loadsheet_preliminary_sent = true end
+    if code == 200 and data.typeL == 2 then toLoadHub.hoppie.loadsheet_chocks_off_sent = true end
+    if code == 200 and data.typeL == 3 then toLoadHub.hoppie.loadsheet_chocks_on_sent = true end
+
     if code ~= 200 then toLoadHub.error_message = "Hoppie returned an error. Please check your secret value." end
     toLoadHub.hoppie.loadsheet_sending = false
 end
@@ -767,6 +800,12 @@ function viewToLoadHubWindow()
                     toLoadHub.next_boarding_check = os.time()
                     toLoadHub.next_cargo_check = os.time()
                     toLoadHub.phases.is_onboarding = true
+                end
+                if imgui.RadioButton("Airstair", not toLoadHub.is_jetbridge) then
+                    toLoadHub.is_jetbridge = false
+                end
+                if imgui.RadioButton("Jetbridge", toLoadHub.is_jetbridge) then
+                    toLoadHub.is_jetbridge = true
                 end
             else
                 imgui.PushStyleColor(imgui.constant.Col.Text, 0xFF95FFF8)
@@ -969,6 +1008,11 @@ function viewToLoadHubWindow()
             imgui.TextUnformatted("All passenger doors will be operated.")
             imgui.PopStyleColor()
             generalSpeed = 2
+        elseif toLoadHub.is_jetbridge then
+            imgui.PushStyleColor(imgui.constant.Col.Text, 0xFF6AE079)
+            imgui.TextUnformatted("Jetbridge attached.")
+            imgui.PopStyleColor()
+            generalSpeed = 2
         end
 
         if isAnyDoorOpen() and (toLoadHub.settings.door.open_boarding and not toLoadHub.phases.is_onboarded) or
@@ -1109,6 +1153,9 @@ function viewToLoadHubWindowSettings()
 
     changed, newval = imgui.Checkbox("Preliminary Loadsheet Only for Long-haul (+7hrs)", toLoadHub.settings.hoppie.preliminary_loadsheet)
     if changed then toLoadHub.settings.hoppie.preliminary_loadsheet , setSave = newval, true end
+
+    changed, newval = imgui.Checkbox("Loadsheet for chocks on and off", toLoadHub.settings.hoppie.chocks_loadsheet)
+    if changed then toLoadHub.settings.hoppie.chocks_loadsheet , setSave = newval, true end
 
     imgui.TextUnformatted("Secret:")
     local masked_secret = string.rep("*", #toLoadHub.settings.hoppie.secret)
@@ -1310,6 +1357,40 @@ function toloadHubMainLoop()
          end
     end
 
+    -- Chocks Off and On
+    if toLoadHub.settings.hoppie.chocks_loadsheet then
+         -- Beacon for Chock Off Loadsheet --
+        if not toLoadHub.chocks_off_set and toLoadHub_beacon_lights_on then
+            toLoadHub.chocks_off_set = true
+            toLoadHub.chocks_off_time = os.date("%H:%M:%S")
+            toLoadHub.hoppie.loadsheet_check = os.time() + 1
+        end
+        -- Engine Off for Chock On Loadsheet --
+        if toLoadHub.hoppie.loadsheet_chocks_off_sent and not toLoadHub.chocks_on_set and isAllEngineOff() then
+            toLoadHub.chocks_on_set = true
+            toLoadHub.chocks_on_time = os.date("%H:%M:%S")
+            toLoadHub.hoppie.loadsheet_check = os.time() + 90
+        end
+
+        -- Altitude Chock Off Loadsheet --
+        if not toLoadHub.hoppie.loadsheet_chocks_off_sent and toLoadHub_pressure_altitude >= 10000 then
+            local data_coff = loadsheetStructure:new()
+            data_coff.typeL = 2
+            data_coff.labelText = "Ch. Off"
+            data_coff.flt_no = toLoadHub_flight_no
+            sendLoadsheetToToliss(data_coff)
+        end
+
+        -- Chock On Loadhseet --
+        if not toLoadHub.hoppie.loadsheet_chocks_on_sent and toLoadHub.chocks_on_set and toLoadHub.hoppie.loadsheet_check < now then
+            local data_con = loadsheetStructure:new()
+            data_con.typeL = 3
+            data_con.labelText = "Ch. On"
+            data_con.flt_no = toLoadHub_flight_no
+            sendLoadsheetToToliss(data_con)
+        end
+    end
+
     -- Focus windows --
     if toLoadHub.focus_windows.pax_load and not toLoadHub.focus_windows.pax_loaded then focusOnToLoadHub() toLoadHub.focus_windows.pax_loaded = true end
     if toLoadHub.focus_windows.cargo_load and not toLoadHub.focus_windows.cargo_loaded then focusOnToLoadHub() toLoadHub.focus_windows.cargo_loaded = true end
@@ -1369,7 +1450,7 @@ function toloadHubMainLoop()
 
     if not toLoadHub.hoppie.loadsheet_sent and toLoadHub.settings.hoppie.enable_loadsheet and toLoadHub.phases.is_onboarded then
         local data_f = loadsheetStructure:new()
-        data_f.isFinal = true
+        data_f.typeL = 1
         data_f.labelText = "@Final@"
         data_f.flt_no = toLoadHub_flight_no
         if toLoadHub_zfw == nil then
@@ -1398,7 +1479,7 @@ function toloadHubMainLoop()
         local data_p = loadsheetStructure:new()
         divideCargoFwdAft()
         setIscsTemporarySimbrief()
-        data_p.isFinal = false
+        data_p.typeL = 0
         data_p.labelText = "Prelim."
         data_p.zfw = string.format("%.1f", toLoadHub.simbrief.est_zfw/1000)
         if toLoadHub_blockZfwCG == nil then
@@ -1450,7 +1531,7 @@ end
 
 if XPLMFindDataRef("toliss_airbus/iscsinterface/zfwCG") then
     dataref("toLoadHub_zfwCG", "toliss_airbus/iscsinterface/zfwCG", "readonly")
-elseif XPLMFindDataRef("toliss_airbus/iscsinterface/zfwCG") then
+elseif XPLMFindDataRef("toliss_airbus/init/ZFWCG") then
     dataref("toLoadHub_zfwCG", "toliss_airbus/init/ZFWCG", "readonly")
 else
     toLoadHub_zfwCG = nil
@@ -1474,6 +1555,9 @@ dataref("toLoadHub_m_fuel_total", "sim/flightmodel/weight/m_fuel_total", "readon
 
 dataref("toLoadHub_flight_no", "toliss_airbus/init/flight_no", "readonly")
 dataref("toLoadHub_WriteFOB_XP", "AirbusFBW/WriteFOB", "readonly")
+
+datared("toLoadHub_pressure_altitude", "sim/flightmodel2/position/pressure_altitude", "readonly")
+datared("toLoadHub_beacon_lights_on", "sim/cockpit/electrical/beacon_lights_on", "readonly")
 
 if XPLMFindDataRef("jd/ghd/execute") then
     dataref("toloadHub_jdexe","jd/ghd/execute", "writeable")